fetchers/linkedin_fetcher.py
from bs4 import BeautifulSoup
from datetime import datetime
import re
from .base_fetcher import BaseFetcher

class LinkedInFetcher(BaseFetcher):
    def __init__(self):
        # In a real implementation, this would be a valid LinkedIn jobs search URL
        super().__init__(site_name="LinkedIn", url="https://www.linkedin.com/jobs/search/")
    
    def parse_jobs(self, html: str):
        """
        Parse LinkedIn job listings from HTML content.
        Returns a list of job dictionaries.
        """
        jobs = []
        soup = BeautifulSoup(html, 'html.parser')
        
        # This is a simplified example - actual LinkedIn parsing would be more complex
        # and would need to adapt to their HTML structure
        job_cards = soup.find_all('div', class_='job-card')
        
        for card in job_cards:
            try:
                # Extract job details
                title_elem = card.find('h3', class_='job-title')
                company_elem = card.find('h4', class_='company-name')
                desc_elem = card.find('p', class_='job-description')
                date_elem = card.find('time', class_='job-date')
                link_elem = card.find('a', class_='job-link')
                
                # Process data
                title = title_elem.text.strip() if title_elem else 'Unknown Title'
                
                # Create description from available elements
                description_parts = []
                if company_elem:
                    description_parts.append(f"Company: {company_elem.text.strip()}")
                if desc_elem:
                    description_parts.append(desc_elem.text.strip())
                description = ' | '.join(description_parts) or 'No description available'
                
                # Get URL
                url = link_elem['href'] if link_elem and 'href' in link_elem.attrs else ''
                
                # Parse date - this is simplified and would need to match LinkedIn's format
                posted_date = None
                if date_elem:
                    date_text = date_elem.text.strip()
                    # Example: convert "2 days ago" to a datetime
                    if 'days ago' in date_text:
                        days = int(re.search(r'(\d+)', date_text).group(1))
                        posted_date = datetime.now().replace(hour=0, minute=0, second=0, microsecond=0)
                        posted_date = posted_date.replace(day=posted_date.day - days)
                    else:
                        posted_date = datetime.now().replace(hour=0, minute=0, second=0, microsecond=0)
                
                # Add job to results
                jobs.append({
                    'title': title,
                    'description': description,
                    'source_site': self.site_name,
                    'url': url,
                    'posted_date': posted_date
                })
                
            except Exception as e:
                # Log error but continue processing other jobs
                print(f"Error parsing job card: {str(e)}")
        
        return jobs