requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.20
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
soupsieve==2.5
typing-extensions==4.7.1
urllib3==2.0.4